# src/main/java/com/adascal/recipesservice/infrastructure/dto/ErrorResponse.java
package com.adascal.recipesservice.infrastructure.dto;

import lombok.AllArgsConstructor;
import lombok.Data;

@AllArgsConstructor
@Data
public class ErrorResponse {
    private String code;
    private String message;
}