# src/main/java/com/adascal/recipesservice/infrastructure/dto/search/SearchRequest.java
package com.adascal.recipesservice.infrastructure.dto.search;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.ArrayList;
import java.util.List;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class SearchRequest {
    private List<SearchQuery> criteria = new ArrayList<>();
}